Mixed performance improvements
- Use SmallRng
- Represent indexes with u16 instead of usize
- Avoid some allocations in montecarlo run

+++ scan_core/src/channel_system/builder.rs
@@ -119,7 +119,7 @@ impl ChannelSystemBuilder {
 
     /// Add a new PG to the CS.
     pub fn new_program_graph(&mut self) -> PgId {
-        let pg_id = PgId(self.program_graphs.len());
+        let pg_id = PgId(self.program_graphs.len() as u16);
         let pg = ProgramGraphBuilder::new();
         self.program_graphs.push(pg);
         pg_id
@@ -131,7 +131,7 @@ impl ChannelSystemBuilder {
     pub fn initial_location(&mut self, pg_id: PgId) -> Result<Location, CsError> {
         let pg = self
             .program_graphs
-            .get(pg_id.0)
+            .get(pg_id.0 as usize)
             .ok_or(CsError::MissingPg(pg_id))?;
         let initial = Location(pg_id, pg.initial_location());
         Ok(initial)
@@ -145,7 +145,7 @@ impl ChannelSystemBuilder {
     pub fn new_var(&mut self, pg_id: PgId, init: CsExpression) -> Result<Var, CsError> {
         let pg = self
             .program_graphs
-            .get_mut(pg_id.0)
+            .get_mut(pg_id.0 as usize)
             .ok_or(CsError::MissingPg(pg_id))?;
         let init = PgExpression::try_from((pg_id, init))?;
         let var = pg
@@ -156,7 +156,7 @@ impl ChannelSystemBuilder {
 
     pub fn new_clock(&mut self, pg_id: PgId) -> Result<Clock, CsError> {
         self.program_graphs
-            .get_mut(pg_id.0)
+            .get_mut(pg_id.0 as usize)
             .ok_or(CsError::MissingPg(pg_id))
             .map(|pg| Clock(pg_id, pg.new_clock()))
     }
@@ -168,7 +168,7 @@ impl ChannelSystemBuilder {
     /// See also [`ProgramGraphBuilder::new_action`].
     pub fn new_action(&mut self, pg_id: PgId) -> Result<Action, CsError> {
         self.program_graphs
-            .get_mut(pg_id.0)
+            .get_mut(pg_id.0 as usize)
             .ok_or(CsError::MissingPg(pg_id))
             .map(|pg| Action(pg_id, pg.new_action()))
     }
@@ -226,7 +226,7 @@ impl ChannelSystemBuilder {
         } else {
             let effect = PgExpression::try_from((pg_id, effect))?;
             self.program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))
                 .and_then(|pg| {
                     pg.add_effect(action.1, var.1, effect)
@@ -247,7 +247,7 @@ impl ChannelSystemBuilder {
             Err(CsError::DifferentPgs(pg_id, clock.0))
         } else {
             self.program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))
                 .and_then(|pg| {
                     pg.reset_clock(action.1, clock.1)
@@ -263,7 +263,7 @@ impl ChannelSystemBuilder {
     /// See also [`ProgramGraphBuilder::new_location`].
     pub fn new_location(&mut self, pg_id: PgId) -> Result<Location, CsError> {
         self.program_graphs
-            .get_mut(pg_id.0)
+            .get_mut(pg_id.0 as usize)
             .ok_or(CsError::MissingPg(pg_id))
             .map(|pg| Location(pg_id, pg.new_location()))
     }
@@ -285,7 +285,7 @@ impl ChannelSystemBuilder {
             })
             .collect::<Result<Vec<_>, CsError>>()?;
         self.program_graphs
-            .get_mut(pg_id.0)
+            .get_mut(pg_id.0 as usize)
             .ok_or(CsError::MissingPg(pg_id))
             .map(|pg| Location(pg_id, pg.new_timed_location(&invariants)))
     }
@@ -315,7 +315,7 @@ impl ChannelSystemBuilder {
                 .map(|guard| PgExpression::try_from((pg_id, guard)))
                 .transpose()?;
             self.program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))
                 .and_then(|pg| {
                     pg.add_transition(pre.1, action.1, post.1, guard)
@@ -355,7 +355,7 @@ impl ChannelSystemBuilder {
                 })
                 .collect::<Result<Vec<_>, CsError>>()?;
             self.program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))
                 .and_then(|pg| {
                     pg.add_timed_transition(pre.1, action.1, post.1, guard, &constraints)
@@ -386,7 +386,7 @@ impl ChannelSystemBuilder {
                 .map(|guard| PgExpression::try_from((pg_id, guard)))
                 .transpose()?;
             self.program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))
                 .and_then(|pg| {
                     pg.add_autonomous_transition(pre.1, post.1, guard)
@@ -423,7 +423,7 @@ impl ChannelSystemBuilder {
                 })
                 .collect::<Result<Vec<_>, CsError>>()?;
             self.program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))
                 .and_then(|pg| {
                     pg.add_autonomous_timed_transition(pre.1, post.1, guard, &constraints)
@@ -437,7 +437,7 @@ impl ChannelSystemBuilder {
     /// - [`None`] capacity means that the channel's capacity is unlimited.
     /// - [`Some(0)`] capacity means the channel uses the handshake protocol (NOT YET IMPLEMENTED!)
     pub fn new_channel(&mut self, var_type: Type, capacity: Option<usize>) -> Channel {
-        let channel = Channel(self.channels.len());
+        let channel = Channel(self.channels.len() as u16);
         self.channels.push((var_type, capacity));
         channel
     }
@@ -453,7 +453,7 @@ impl ChannelSystemBuilder {
     ) -> Result<Action, CsError> {
         let channel_type = self
             .channels
-            .get(channel.0)
+            .get(channel.0 as usize)
             .ok_or(CsError::MissingChannel(channel))?
             .0
             .to_owned();
@@ -462,7 +462,7 @@ impl ChannelSystemBuilder {
         if channel_type != message_type {
             Err(CsError::ProgramGraph(pg_id, PgError::TypeMismatch))
         } else {
-            let action = self.program_graphs[pg_id.0]
+            let action = self.program_graphs[pg_id.0 as usize]
                 .new_send(msg)
                 .map_err(|err| CsError::ProgramGraph(pg_id, err))?;
             let action = Action(pg_id, action);
@@ -485,13 +485,13 @@ impl ChannelSystemBuilder {
         } else {
             let channel_type = self
                 .channels
-                .get(channel.0)
+                .get(channel.0 as usize)
                 .ok_or(CsError::MissingChannel(channel))?
                 .0
                 .to_owned();
             let message_type = self
                 .program_graphs
-                .get(pg_id.0)
+                .get(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))?
                 .var_type(var.1)
                 .map_err(|err| CsError::ProgramGraph(pg_id, err))?
@@ -499,7 +499,7 @@ impl ChannelSystemBuilder {
             if channel_type != message_type {
                 Err(CsError::ProgramGraph(pg_id, PgError::TypeMismatch))
             } else {
-                let action = self.program_graphs[pg_id.0]
+                let action = self.program_graphs[pg_id.0 as usize]
                     .new_receive(var.1)
                     .map_err(|err| CsError::ProgramGraph(pg_id, err))?;
                 let action = Action(pg_id, action);
@@ -520,7 +520,7 @@ impl ChannelSystemBuilder {
     ) -> Result<Action, CsError> {
         let (_, cap) = self
             .channels
-            .get(channel.0)
+            .get(channel.0 as usize)
             .ok_or(CsError::MissingChannel(channel))?;
         if matches!(cap, Some(0)) {
             // it makes no sense to probe an handshake channel
@@ -528,7 +528,7 @@ impl ChannelSystemBuilder {
         } else {
             let action = self
                 .program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))?
                 .new_action();
             let action = Action(pg_id, action);
@@ -548,7 +548,7 @@ impl ChannelSystemBuilder {
     ) -> Result<Action, CsError> {
         let (_, cap) = self
             .channels
-            .get(channel.0)
+            .get(channel.0 as usize)
             .ok_or(CsError::MissingChannel(channel))?;
         if matches!(cap, Some(0)) {
             // it makes no sense to probe an handshake channel
@@ -559,7 +559,7 @@ impl ChannelSystemBuilder {
         } else {
             let action = self
                 .program_graphs
-                .get_mut(pg_id.0)
+                .get_mut(pg_id.0 as usize)
                 .ok_or(CsError::MissingPg(pg_id))?
                 .new_action();
             let action = Action(pg_id, action);
@@ -589,9 +589,9 @@ impl ChannelSystemBuilder {
         let mut communications = Vec::with_capacity(communications_map.len());
         let mut communications_pg_idxs = Vec::with_capacity(program_graphs.len() + 1);
         communications_pg_idxs.push(0);
-        let mut current_pg = 0usize;
-        let mut counter = 0usize;
-        let mut prev_counter = 0usize;
+        let mut current_pg = 0u16;
+        let mut counter = 0u16;
+        let mut prev_counter = 0u16;
         for (action, (c, m)) in communications_map.into_iter() {
             communications.push((action.1, c, m));
             let pg_id = action.0;
@@ -605,7 +605,7 @@ impl ChannelSystemBuilder {
             counter += 1;
         }
         communications_pg_idxs
-            .extend((0..(program_graphs.len() - 1 - current_pg)).map(|_| prev_counter));
+            .extend((0..(program_graphs.len() as u16 - 1 - current_pg)).map(|_| prev_counter));
         communications_pg_idxs.push(counter);
         assert_eq!(communications_pg_idxs.len(), program_graphs.len() + 1);
